split posture module
